+++ src/Components/Animation/index.tsx
@@ -1,5 +1,7 @@
+import { FadeOut } from './FadeOut';
 import { SlideFromLeft } from './SlideFromLeft';
 
 export const Animation = {
   SlideFromLeft: SlideFromLeft,
+  FadeOut: FadeOut,
 };