+++ docs/ReleaseNotes.md
@@ -1,6 +1,6 @@
 // Before any commit, use the command line: npx tsc --noEmit
 
-# 0.1.9.22 Image unavailable handling.
+# 0.1.9.23 Image Description Input
 
-- Now if a image is not available on user phone, a button will appear in place, to allow the user to download the image from a server. Donwload endpoint not implemtede yet
-- Now the app keeps tracking of media files on cache, to decide if the media download button should appear or not.
+- Input for image description implemented.
+- Added a <Animation.FadeOut /> component